requirements.txt
Django==4.1.3
python-dotenv==0.21.0
dj-database-url==0.5.0
Pillow==9.0.0
gunicorn==20.1.0
django-extensions==3.2.0
drf_yasg==1.21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-widget-tweaks==1.4.12
django-crispy-forms==1.14.0
django-ckeditor==6.5.1
django-taggit==3.1.0